refactory: rm unnecessary clones & rename some files
- rm unnecessary clones of `ColStats`
- rename fuse table operation mods
