+++ query/src/datasources/table/fuse/append.rs
