+++ query/src/datasources/table/fuse/table.rs
@@ -14,8 +14,6 @@
 //
 
 use std::any::Any;
-use std::collections::HashSet;
-use std::iter::FromIterator;
 use std::sync::Arc;
 
 use common_context::DataContext;
@@ -26,7 +24,6 @@ use common_exception::Result;
 use common_meta_types::TableInfo;
 use common_planners::Extras;
 use common_planners::InsertIntoPlan;
-use common_planners::Part;
 use common_planners::Partitions;
 use common_planners::ReadDataSourcePlan;
 use common_planners::Statistics;
@@ -35,7 +32,6 @@ use common_streams::SendableDataBlockStream;
 
 use super::util;
 use crate::catalogs::Table;
-use crate::datasources::table::fuse::BlockMeta;
 use crate::datasources::table::fuse::TableSnapshot;
 
 pub struct FuseTable {
@@ -121,38 +117,5 @@ impl FuseTable {
         } else {
             Ok(None)
         }
-    }
-
-    pub(crate) fn to_partitions(
-        blocks_metas: &[BlockMeta],
-        push_downs: Option<Extras>,
-    ) -> (Statistics, Partitions) {
-        let proj_cols =
-            push_downs.and_then(|extras| extras.projection.map(HashSet::<usize>::from_iter));
-        blocks_metas.iter().fold(
-            (Statistics::default(), Partitions::default()),
-            |(mut stats, mut parts), block_meta| {
-                parts.push(Part {
-                    name: block_meta.location.location.clone(),
-                    version: 0,
-                });
-
-                stats.read_rows += block_meta.row_count as usize;
-
-                match &proj_cols {
-                    Some(proj) => {
-                        stats.read_bytes += block_meta
-                            .col_stats
-                            .iter()
-                            .filter(|(cid, _)| proj.contains(&(**cid as usize)))
-                            .map(|(_, col_stats)| col_stats.in_memory_size)
-                            .sum::<u64>() as usize
-                    }
-                    None => stats.read_bytes += block_meta.block_size as usize,
-                }
-
-                (stats, parts)
-            },
-        )
     }
 }
