+++ query/src/datasources/table/fuse/table_test.rs
@@ -13,13 +13,10 @@
 //  limitations under the License.
 //
 
-use std::collections::HashMap;
 use std::sync::Arc;
 
 use common_base::tokio;
-use common_datavalues::DataValue;
 use common_exception::Result;
-use common_planners::Extras;
 use common_planners::ReadDataSourcePlan;
 use common_planners::TruncateTablePlan;
 use futures::TryStreamExt;
@@ -27,10 +24,6 @@ use futures::TryStreamExt;
 use crate::catalogs::Catalog;
 use crate::catalogs::ToReadDataSourcePlan;
 use crate::datasources::table::fuse::table_test_fixture::TestFixture;
-use crate::datasources::table::fuse::BlockLocation;
-use crate::datasources::table::fuse::BlockMeta;
-use crate::datasources::table::fuse::ColStats;
-use crate::datasources::table::fuse::FuseTable;
 
 #[tokio::test]
 async fn test_fuse_table_simple_case() -> Result<()> {
@@ -200,75 +193,5 @@ async fn test_fuse_table_truncate() -> Result<()> {
     assert_eq!(parts.len(), 0);
     assert_eq!(stats.read_rows, 0);
 
-    Ok(())
-}
-
-#[test]
-fn test_fuse_table_to_partitions_stats_with_col_pruning() -> Result<()> {
-    // setup
-    let num_of_col = 10;
-    let num_of_block = 5;
-
-    let col_stats_gen = |col_size| ColStats {
-        min: DataValue::Int8(Some(1)),
-        max: DataValue::Int8(Some(2)),
-        null_count: 0,
-        in_memory_size: col_size as u64,
-    };
-
-    let cols_stats = (0..num_of_col)
-        .into_iter()
-        .map(|col_id| (col_id as u32, col_stats_gen(col_id)))
-        .collect::<HashMap<_, _>>();
-
-    let block_meta = BlockMeta {
-        row_count: 0,
-        block_size: cols_stats
-            .iter()
-            .map(|(_, col_stats)| col_stats.in_memory_size)
-            .sum(),
-        col_stats: cols_stats.clone(),
-        location: BlockLocation {
-            location: "".to_string(),
-            meta_size: 0,
-        },
-    };
-
-    let blocks_metas = (0..num_of_block)
-        .into_iter()
-        .map(|_| block_meta.clone())
-        .collect::<Vec<_>>();
-
-    // CASE I:  no projection
-    let (s, _) = FuseTable::to_partitions(&blocks_metas, None);
-    let expected_block_size: u64 = cols_stats
-        .iter()
-        .map(|(_, col_stats)| col_stats.in_memory_size)
-        .sum();
-    assert_eq!(expected_block_size * num_of_block, s.read_bytes as u64);
-
-    // CASE II: col pruning
-    // projection which keeps the odd ones
-    let proj = (0..num_of_col)
-        .into_iter()
-        .filter(|v| v & 1 != 0)
-        .collect::<Vec<usize>>();
-
-    // for each block, the block size we expects (after pruning)
-    let expected_block_size: u64 = cols_stats
-        .iter()
-        .filter(|(cid, _)| proj.contains(&(**cid as usize)))
-        .map(|(_, col_stats)| col_stats.in_memory_size)
-        .sum();
-
-    // kick off
-    let push_down = Some(Extras {
-        projection: Some(proj),
-        filters: vec![],
-        limit: None,
-        order_by: vec![],
-    });
-    let (stats, _) = FuseTable::to_partitions(&blocks_metas, push_down);
-    assert_eq!(expected_block_size * num_of_block, stats.read_bytes as u64);
     Ok(())
 }
