+++ query/src/datasources/table/fuse/truncate.rs
